+++ rt-trace/src/command.rs
@@ -1,0 +1,5 @@
+use crate::span::RawSpan;
+
+pub(crate) enum Command {
+    SendSpans(Vec<RawSpan>),
+}
